# Conflicts:
#	package.json
#	src/chromeDebugAdapter.ts

+++ src/chromeDebugAdapter.ts
@@ -2,18 +2,22 @@
  * Copyright (C) Microsoft Corporation. All rights reserved.
  *--------------------------------------------------------*/
 
-import {ChromeDebugAdapter as CoreDebugAdapter, logger, utils as coreUtils, ISourceMapPathOverrides, ICommonRequestArgs} from 'vscode-chrome-debug-core';
-import {spawn, ChildProcess} from 'child_process';
-import Crdp from 'chrome-remote-debug-protocol';
+import * as os from 'os';
+import * as path from 'path';
+
+import {ChromeDebugAdapter as CoreDebugAdapter, logger, utils as coreUtils, ISourceMapPathOverrides, stoppedEvent} from 'vscode-chrome-debug-core';
+import {spawn, ChildProcess, fork, execSync} from 'child_process';
+import {Crdp} from 'vscode-chrome-debug-core';
 import {DebugProtocol} from 'vscode-debugprotocol';
 
-import {ILaunchRequestArgs, IAttachRequestArgs} from './chromeDebugInterfaces';
+import {ILaunchRequestArgs, IAttachRequestArgs, ICommonRequestArgs} from './chromeDebugInterfaces';
 import * as utils from './utils';
 
 import * as os from 'os';
 import * as path from 'path';
 
 const DefaultWebSourceMapPathOverrides: ISourceMapPathOverrides = {
+    'webpack:///./~/*': '${webRoot}/node_modules/*',
     'webpack:///./*': '${webRoot}/*',
     'webpack:///*': '*',
     'meteor://💻app/*': '${webRoot}/*',
@@ -44,6 +48,7 @@ export class ChromeDebugAdapter extends CoreDebugAdapter {
 
     private _chromeProc: ChildProcess;
     private _overlayHelper: utils.DebounceHelper;
+    private _chromePID: number;
 
     private _kha: string;
 
@@ -113,12 +118,7 @@ export class ChromeDebugAdapter extends CoreDebugAdapter {
                 const kromArgs: string[] = [path.join(args.cwd, 'build', 'krom'), path.join(args.cwd, 'build', 'krom-resources'), '--debug', port.toString(), '--watch'];
 
                 logger.log(`spawn('${kromPath}', ${JSON.stringify(kromArgs) })`);
-                this._chromeProc = spawn(kromPath, kromArgs, {
-                    detached: true,
-                    stdio: ['ignore'],
-                    cwd: path.join(args.krom, osDir())
-                });
-                this._chromeProc.unref();
+                this._chromeProc = spawnChrome(kromPath, kromArgs, !!args.runtimeExecutable);
                 this._chromeProc.on('error', (err) => {
                     const errMsg = 'Krom error: ' + err;
                     logger.error(errMsg);
@@ -128,7 +128,8 @@ export class ChromeDebugAdapter extends CoreDebugAdapter {
                 /*return new Promise<void>((resolve, reject) => {
                     resolve();
                 });*/
-                return this.doAttach(port, 'http://krom', args.address);
+                return args.noDebug ? undefined :
+                    this.doAttach(port, 'http://krom', args.address, args.timeout);
             }, (reason) => {
                 logger.error('Launch canceled.', true);
                 require(path.join(this._kha, 'Tools/khamake/out/main.js')).close();
@@ -140,10 +141,20 @@ export class ChromeDebugAdapter extends CoreDebugAdapter {
     }
 
     public attach(args: IAttachRequestArgs): Promise<void> {
+        if (args.urlFilter) {
+            args.url = args.urlFilter;
+        }
+
         return super.attach(args);
     }
 
     public commonArgs(args: ICommonRequestArgs): void {
+        if (!args.webRoot && args.pathMapping && args.pathMapping['/']) {
+            // Adapt pathMapping['/'] as the webRoot when not set, since webRoot is explicitly used in many places
+            args.webRoot = args.pathMapping['/'];
+        }
+
+        args.sourceMaps = typeof args.sourceMaps === 'undefined' || args.sourceMaps;
         args.sourceMapPathOverrides = getSourceMapPathOverrides(args.webRoot, args.sourceMapPathOverrides);
         args.skipFileRegExps = ['^chrome-extension:.*'];
 
@@ -159,16 +170,20 @@ export class ChromeDebugAdapter extends CoreDebugAdapter {
     }
 
     protected runConnection(): Promise<void>[] {
-        return [...super.runConnection()];//, this.chrome.Page.enable()];
+        return [...super.runConnection()];//, this.chrome.Page.enable(), this.chrome.Network.enable({})];
     }
 
     protected onEntryAdded(event: Crdp.Log.EntryAddedEvent): void {
         logger.log(event.entry.text, true);
     }
 
-    protected onPaused(notification: Crdp.Debugger.PausedEvent): void {
+    protected onPaused(notification: Crdp.Debugger.PausedEvent, expectingStopReason?: stoppedEvent.ReasonType): void {
         this._overlayHelper.doAndCancel(() => this.chrome.Page.configureOverlay({ message: ChromeDebugAdapter.PAGE_PAUSE_MESSAGE }).catch(() => { }));
-        super.onPaused(notification);
+        super.onPaused(notification, expectingStopReason);
+    }
+
+    protected threadName(): string {
+        return 'Chrome';
     }
 
     protected onResumed(): void {
@@ -177,14 +192,33 @@ export class ChromeDebugAdapter extends CoreDebugAdapter {
     }
 
     public disconnect(): void {
-        if (this._chromeProc) {
-            this._chromeProc.kill('SIGINT');
-            this._chromeProc = null;
+        const hadTerminated = this._hasTerminated;
+
+        // Disconnect before killing Chrome, because running "taskkill" when it's paused sometimes doesn't kill it
+        super.disconnect();
+
+        if (this._chromeProc && !hadTerminated) {
+            // Only kill Chrome if the 'disconnect' originated from vscode. If we previously terminated
+            // due to Chrome shutting down, or devtools taking over, don't kill Chrome.
+            if (coreUtils.getPlatform() === coreUtils.Platform.Windows && this._chromePID) {
+                // Run synchronously because this process may be killed before exec() would run
+                const taskkillCmd = `taskkill /F /T /PID ${this._chromePID}`;
+                logger.log(`Killing Chrome process by pid: ${taskkillCmd}`);
+                try {
+                    execSync(taskkillCmd);
+                } catch (e) {
+                    // Can fail if Chrome was already open, and the process with _chromePID is gone.
+                    // Or if it already shut down for some reason.
+                }
+            } else {
+                logger.log('Killing Chrome process');
+                this._chromeProc.kill('SIGINT');
+            }
         }
 
         require(path.join(this._kha, 'Tools/khamake/out/main.js')).close();
 
-        return super.disconnect();
+        this._chromeProc = null;
     }
 
     public runScript(): void {
@@ -199,6 +233,42 @@ export class ChromeDebugAdapter extends CoreDebugAdapter {
      */
     public restart(): Promise<void> {
         return this.chrome.Page.reload({ ignoreCache: true });
+    }
+
+    private spawnChrome(chromePath: string, chromeArgs: string[], usingRuntimeExecutable: boolean): ChildProcess {
+        if (coreUtils.getPlatform() === coreUtils.Platform.Windows && !usingRuntimeExecutable) {
+            const chromeProc = fork(getChromeSpawnHelperPath(), [chromePath, ...chromeArgs], { execArgv: [], silent: true });
+            chromeProc.unref();
+
+            chromeProc.on('message', data => {
+                const pidStr = data.toString();
+                logger.log('got chrome PID: ' + pidStr);
+                this._chromePID = parseInt(pidStr, 10);
+            });
+
+            chromeProc.on('error', (err) => {
+                const errMsg = 'chromeSpawnHelper error: ' + err;
+                logger.error(errMsg);
+            });
+
+            chromeProc.stderr.on('data', data => {
+                logger.error('[chromeSpawnHelper] ' + data.toString());
+            });
+
+            chromeProc.stdout.on('data', data => {
+                logger.log('[chromeSpawnHelper] ' + data.toString());
+            });
+
+            return chromeProc;
+        } else {
+            logger.log(`spawn('${chromePath}', ${JSON.stringify(chromeArgs) })`);
+            const chromeProc = spawn(chromePath, chromeArgs, {
+                detached: true,
+                stdio: ['ignore'],
+            });
+            chromeProc.unref();
+            return chromeProc;
+        }
     }
 }
 
@@ -229,4 +299,13 @@ export function resolveWebRootPattern(webRoot: string, sourceMapPathOverrides: I
     }
 
     return resolvedOverrides;
+}
+
+function getChromeSpawnHelperPath(): string {
+    if (path.basename(__dirname) === 'src') {
+        // For tests
+        return path.join(__dirname, '../chromeSpawnHelper.js');
+    } else {
+        return path.join(__dirname, 'chromeSpawnHelper.js');
+    }
 }
